+++ pyogrio/core.py
@@ -178,6 +178,7 @@ def read_info(path_or_buffer, /, layer=None, encoding=None, **kwargs):
                 "geometry": "<geometry type>",
                 "features": <feature count>,
                 "driver": "<driver>",
+                "metadata" "<dict of layer metadata or None>"
             }
     """
     path, buffer = get_vsi_path(path_or_buffer)