+++ pyogrio/geopandas.py
@@ -204,6 +204,7 @@ def write_dataframe(
     promote_to_multi=None,
     nan_as_null=True,
     append=False,
+    metadata=None,
     dataset_options=None,
     layer_options=None,
     **kwargs,
@@ -261,6 +262,10 @@ def write_dataframe(
         driver supports appending to an existing data source, will cause the
         data to be appended to the existing records in the data source.
         NOTE: append support is limited to specific drivers and GDAL versions.
+    metadata : dict, optional (defaul: None)
+        Metadata to be stored in the output file; limited to drivers that
+        support writing metadata (GPKG) and silently ignored otherwise.  Keys
+        and values must be strings.
     dataset_options : dict, optional
         Dataset creation option (format specific) passed to OGR. Specify as
         a key-value dictionary.
@@ -393,6 +398,7 @@ def write_dataframe(
         promote_to_multi=promote_to_multi,
         nan_as_null=nan_as_null,
         append=append,
+        metadata=metadata,
         dataset_options=dataset_options,
         layer_options=layer_options,
         **kwargs,