+++ CHANGES.md
@@ -10,6 +10,9 @@
     `read_info` (#233)
 -   Standardized 3-dimensional geometry type labels from "2.5D <type>" to
     "<type> Z" for consistency with well-known text (WKT) formats (#234)
+-   Add support for writing dataset metadata (where supported by driver) to
+    `write` and `write_dataframe`, and add support for reading dataset metadata
+    in `read_info`.
 
 ## 0.5.1 (2023-01-26)